+++ lib/zine/watcher.rb
@@ -5,6 +5,7 @@ module Zine
   # Watch files for changes
   class Watcher
     attr_reader :upload_array, :delete_array
+    attr_accessor :listener_array
 
     def initialize(posts_and_headlines, build_directory, source_directory)
       @posts_and_headlines = posts_and_headlines
@@ -12,6 +13,7 @@ module Zine
       @source_directory = File.join Dir.pwd, source_directory
       @upload_array = []
       @delete_array = []
+      @listener_array = []
     end
 
     # Build a delete list & an upload list for SSH from changes in build,
@@ -74,6 +76,7 @@ module Zine
         on_build_delete removed unless removed.empty?
       end
       listener.start
+      @listener_array << listener
     end
 
     def watch_source_dir
@@ -83,6 +86,7 @@ module Zine
         on_source_delete removed unless removed.empty?
       end
       listener.start
+      @listener_array << listener
     end
   end
 end